requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.5
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.1
fastapi==0.115.14
filelock==3.18.0
filetype==1.2.0
fsspec==2025.5.1
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
gTTS==2.5.4
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.2
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.4
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
oauthlib==3.3.1
olefile==0.47
openai==1.93.0
orjson==3.10.18
packaging==24.2
pillow==11.3.0
portalocker==2.10.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pypdf==5.7.0
pytesseract==0.3.13
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.3
RapidFuzz==3.13.0
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
spotipy==2.25.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
thefuzz==0.22.1
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
unstructured==0.18.3
unstructured-client==0.38.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
webencodings==0.5.1
wrapt==1.17.2
yt-dlp==2025.6.30
zstandard==0.23.0
sqlalchemy "psycopg2-binary" pydantic-settings
alembic
pydantic[email]
python-jose[cryptography]